requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pdf2image>=1.17
pillow>=10.3
openai>=1.30
PyMuPDF>=1.24
python-multipart>=0.0.9